\name{whale}
\alias{whale}
\docType{data}
\title{Whale data}
\description{
  Nursing times for a baby beluga whale.
}
\usage{data(whale)}
\format{
  A data frame with 228 observations on the following 3 variables.
  \describe{
    \item{V1}{a numeric vector}
    \item{V2}{a numeric vector}
    \item{V3}{a factor with levels \code{0} \code{104} \code{118} \code{119} \code{126} \code{127} \code{132} \code{135} \code{137} \code{14} \code{144} \code{146} \code{150} \code{151} \code{153} \code{156} \code{157} \code{160} \code{166} \code{167} \code{168} \code{169} \code{170} \code{171} \code{172} \code{174} \code{175} \code{176} \code{180} \code{186} \code{187} \code{189} \code{191} \code{192} \code{193} \code{196} \code{197} \code{198} \code{199} \code{200} \code{204} \code{205} \code{216} \code{218} \code{222} \code{223} \code{225} \code{226} \code{228} \code{229} \code{230} \code{231} \code{232} \code{236} \code{239} \code{243} \code{244} \code{247} \code{252} \code{253} \code{255} \code{257} \code{260} \code{267} \code{271} \code{274} \code{275} \code{277} \code{284} \code{285} \code{286} \code{288} \code{291} \code{292} \code{299} \code{308} \code{320} \code{323} \code{326} \code{332} \code{338} \code{339} \code{340} \code{344} \code{345} \code{349} \code{351} \code{353} \code{354} \code{359} \code{360} \code{362} \code{371} \code{372} \code{377} \code{380} \code{386} \code{404} \code{409} \code{411} \code{419} \code{423} \code{426} \code{429} \code{430} \code{432} \code{433} \code{435} \code{438} \code{440} \code{441} \code{442} \code{443} \code{444} \code{445} \code{446} \code{449} \code{450} \code{453} \code{456} \code{462} \code{463} \code{464} \code{470} \code{473} \code{477} \code{48} \code{485} \code{491} \code{492} \code{494} \code{495} \code{497} \code{504} \code{506} \code{509} \code{51} \code{513} \code{515} \code{524} \code{528} \code{533} \code{537} \code{538} \code{541} \code{565} \code{579} \code{59} \code{590} \code{600} \code{605} \code{613} \code{644} \code{648} \code{659} \code{68} \code{688} \code{69} \code{693} \code{694} \code{702} \code{714} \code{72} \code{720} \code{737} \code{74} \code{750} \code{756} \code{772} \code{80} \code{805} \code{813} \code{825} \code{84} \code{85} \code{870} \code{873} \code{888} \code{92} \code{93} \code{954} \code{96} \code{98} \code{M}}
  }
}
\source{
  Simonoff, J. Smoothing Methods in Statistics, Springer, 1996.
}
\keyword{datasets}
